FFC-148185 renamed schema folder
