+++ app/auto-eligibility/schemas/sbi.schema.js
